Iris 1.2-beta2
 + Sherpa is launched and killed by the launching script, not by Specview.
 + The launching script sends a SIGINT signal to sherpa, so that it can shut down cleanly.
 + The Plugin Framework has been completely redesigned and it should be now ready for production. In particular, the new framework allows to install and uninstall plugins "on the fly", and supports custom icons.
 + The new Plugin Manager is the GUI for installing and uninstalling additional plugins. You can both install and uninstall them at runtime... no need to restart Iris. The additional buttons and menus will come and go.
 + An example Plugin is included in the examples directory, to exercise the Plugin Manager.
 + The SDK itself is almost ready: a maven archetype with the example plugin and the ability to include classes from external libraries and customized icons. Some details have to be ironed out, but they have to do with the SDK itself, not with the main Iris Development.

Jira Tickets addressed:
VAOPD-667 Start Sherpa from the Iris launcher (not from the Visualizer component)
VAOPD-612 Adding a "Duplicate Photometric Point" button in the IRIS builder Photometric Catalog Importer (However, after an iteration with Raffaele we decided not to create a new button, but that when a new point is created it inherits the properties of the last point created. This should be more user friendly).
BR-117 SED Builder fails to show error message when an incompatible spectrum/SED is added